+++ internal/server/server.go
@@ -42,7 +42,7 @@ import (
 )
 
 var (
-	Version     = "1.8.1"
+	Version     = "1.8.2"
 	connCounter uint64
 )